Update patch set 1
Patch Set 1: Code-Review-2

We won't be accepting these, sorry.  At one point in time, we used to post process all the PNGs in the generated update zip, but storage on most devices has gotten to the point where the additional 5 minutes of build time is not worth it.  Preprocessing the PNGs is no-go either, since this will just cause merge conflicts when we pull from Google next time.

Patch-set: 1
Label: Code-Review=-2